Fixed problem whereby Cablecast module was stripping existing taxonomy from projects.
I created a function called "cablecast_merge_terms", which checks to see if there are other terms 
already on the node in this vocabulary, and if there are, and if the vocabulary is not freetagging, 
then removes the preexisting term. Then it adds the term. All other terms on the node are preserved.

I applied this method in cablecast_sync_shows and cablecast_sync_projects.